Refactored mocks for Pipeline Utility Steps Plugin
